backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pdfplumber==0.11.2
PyPDF2==3.0.1
python-docx==1.1.2
openai==1.40.0
gunicorn==22.0.0
